Add files via upload
Here is the folder for my sprint 2 work. Within the folder you'll find html, js, and css files along with some images and mp3s. 
Popup-sound.html: creates content for the wii inspired popup
popup-sound.css: styles the wii popup
window-manipulator.css,html, and js: This styles a column on the left side of the screen that has an externally coded function for window manipulation. This is where i took the code and tried to modify it to my own.